+++ py/prospect/scripts/__init__.py
@@ -1,8 +1,9 @@
-#
-# See top-level LICENSE.rst file for Copyright information
-#
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
 # -*- coding: utf-8 -*-
+"""
+================
+prospect.scripts
+================
 
-# help with 2to3 support.
-from __future__ import absolute_import, division, print_function
-
+Code to support command-line scripts.
+"""