+++ py/prospect/scripts/specview_per_pixel.py
@@ -1,8 +1,11 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+# -*- coding: utf-8 -*-
 """
+===================================
 prospect.scripts.specview_per_pixel
 ===================================
 
-Write static html files from coadded spectra, sorted by healpixels
+Write static html files from coadded spectra, sorted by healpixels.
 """
 
 import os, sys, glob
@@ -20,11 +23,12 @@ from desitarget.sv1.sv1_targetmask import desi_mask as sv1_desi_mask
 import desispec.spectra
 import desispec.frame
 
-from prospect import myspecselect # special (to be edited)
-from prospect import plotframes
-from prospect import utils_specviewer
+from ..myspecselect import myspecselect  # special (to be edited)
+from ..plotframes import coadd_targets, create_model, plotspectra
+from ..utilities import specviewer_selection, match_zcat_to_spectra, match_vi_targets  #, miniplot_spectrum
 
-def parse() :
+
+def _parse():
 
     parser = argparse.ArgumentParser(description='Create pixel-based static html pages for the spectral viewer')
     parser.add_argument('--specprod_dir', help='overrides $DESI_SPECTRO_REDUX/$SPECPROD/', type=str, default=None)
@@ -43,14 +47,14 @@ def parse() :
     return args
 
 
-def main(args) :
-    
+def main():
+    args = _parse()
     log = get_logger()
     specprod_dir = args.specprod_dir
     if specprod_dir is None : specprod_dir = desispec.io.specprod_root()
     webdir = args.webdir
     if webdir is None : webdir = os.environ["DESI_WWW"]+"/users/armengau/svdc2019c" # TMP, for test
-            
+
     # TODO - Selection on pixels based on existing specviewer pages
     if args.pixel_list is None :
         pixels = glob.glob( os.path.join(specprod_dir,"spectra-64/*/*") )
@@ -59,24 +63,24 @@ def main(args) :
         pixels = np.loadtxt(args.pixel_list, dtype=str)
     if args.random_pixels :
         random.shuffle(pixels)
-        
+
     # Loop on pixels
     nspec_done = 0
     for pixel in pixels :
-        
+
         log.info("Working on pixel "+pixel)
         thefile = desispec.io.findfile('spectra', groupname=int(pixel), specprod_dir=specprod_dir)
         individual_spectra = desispec.io.read_spectra(thefile)
-        spectra = plotframes.coadd_targets(individual_spectra)
+        spectra = coadd_targets(individual_spectra)
         zbfile = thefile.replace('spectra-64-', 'zbest-64-')
         if os.path.isfile(zbfile) :
             zbest = Table.read(zbfile, 'ZBEST')
         else :
             log.info("No associated zbest file found : skipping pixel")
             continue
-        
-        spectra = utils_specviewer.specviewer_selection(spectra, log=log,
-                        mask=args.mask, mask_type=args.mask_type, gmag_cut=args.gcut, rmag_cut=args.rcut, 
+
+        spectra = specviewer_selection(spectra, log=log,
+                        mask=args.mask, mask_type=args.mask_type, gmag_cut=args.gcut, rmag_cut=args.rcut,
                         chi2cut=args.chi2cut, zbest=zbest)
         if spectra == 0 : continue
 
@@ -84,17 +88,17 @@ def main(args) :
         # TODO - Find a more useful sort ?
         nbpages = int(np.ceil((spectra.num_spectra()/args.nspecperfile)))
         sort_indices = np.argsort(spectra.fibermap["TARGETID"])
-        
+
         for i_page in range(1,1+nbpages) :
-            
+
             log.info(" * Page "+str(i_page)+" / "+str(nbpages))
             the_indices = sort_indices[(i_page-1)*args.nspecperfile:i_page*args.nspecperfile]
-            thespec = myspecselect.myspecselect(spectra, indices=the_indices)
-            thezb, kk = utils_specviewer.match_zcat_to_spectra(zbest,thespec)
+            thespec = myspecselect(spectra, indices=the_indices)
+            thezb, kk = match_zcat_to_spectra(zbest,thespec)
             ### No VI results to display by default
             # VI "catalog" - location to define later ..
             # vifile = os.environ['HOME']+"/prospect/vilist_prototype.fits"
-            # vidata = utils_specviewer.match_vi_targets(vifile, thespec.fibermap["TARGETID"])
+            # vidata = match_vi_targets(vifile, thespec.fibermap["TARGETID"])
             titlepage = "pix"+pixel+"_"+str(i_page)
             if args.gcut is not None :
                 titlepage = "gcut-"+str(args.gcut[0])+"-"+str(args.gcut[1])+"_"+titlepage
@@ -104,22 +108,20 @@ def main(args) :
                 titlepage = "chi2cut-"+str(args.chi2cut[0])+"-"+str(args.chi2cut[1])+"_"+titlepage
             if args.mask is not None :
                 titlepage = args.mask+"_"+titlepage
-            model = plotframes.create_model(thespec, thezb)
+            model = create_model(thespec, thezb)
             html_dir = os.path.join(webdir,"pix"+pixel)
-            if not os.path.exists(html_dir) : 
+            if not os.path.exists(html_dir) :
                 os.makedirs(html_dir)
                 os.mkdir(html_dir+"/vignettes")
-            
-            plotframes.plotspectra(thespec, zcatalog=zbest, model_from_zcat=True, vidata=None, model=None, title=titlepage, html_dir=html_dir, is_coadded=True, mask_type=args.mask_type)
+
+            plotspectra(thespec, zcatalog=zbest, model_from_zcat=True, vidata=None, model=None, title=titlepage, html_dir=html_dir, is_coadded=True, mask_type=args.mask_type)
 #             for i_spec in range(thespec.num_spectra()) :
 #                 saveplot = html_dir+"/vignettes/pix"+pixel+"_"+str(i_page)+"_"+str(i_spec)+".png"
-#                 utils_specviewer.miniplot_spectrum(thespec, i_spec, model=model, saveplot=saveplot, smoothing = args.vignette_smoothing)
+#                 miniplot_spectrum(thespec, i_spec, model=model, saveplot=saveplot, smoothing = args.vignette_smoothing)
             nspec_done += thespec.num_spectra()
-        
+
         # Stop running if needed, only once a full pixel is completed
         if args.nmax_spectra is not None :
             if nspec_done >= args.nmax_spectra :
                 log.info(str(nspec_done)+" spectra done : no other pixel will be processed")
                 break
-
-